Add commands to read fuse settings, for SMC diagnostics (14 bytes)
New size: 7028 (max size 7680)

Reading fuse bits (and lock bits) is a hardware feature where you set
SPMCSR to (1 << RFLB) | (1 << SELFPRGEN), and then LPM within 4 clock
cycles.

The feature you read with LPM is decided by Z:
0: Low fuse
1: Lock bits
2: Extended fuse
3: High fuse

To keep firmware use at a minimum, assume that the numeric values of the
I2C commands are in the same order as the hardware feature, so that they
can share the code, and only do a simple subtract to separate between them.

X16 need to know which fuse settings to expect, and which fuse settings are
incorrect.